backend/pyproject.toml
[tool.poetry]
name = "history-ai-wiki-backend"
version = "0.1.0"
description = "Backend for History AI Wiki using FastAPI"
packages = [{ include = "app" }]

[tool.poetry.dependencies]
python = ">=3.11"
fastapi = ">=0.115.0"
uvicorn = {extras = ["standard"], version = ">=0.30.0"}
pydantic = ">=2.8.0"
python-dotenv = ">=1.0.1"
sqlalchemy = ">=2.0.28"
psycopg = {version = ">=3.2.0", extras = ["binary"]}
pytest = "^8.4.2"
google-generativeai = "^0.8.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
